fastapi/pyproject.toml
[tool.poetry]
name = "chatbot-server"
version = "0.1.0"
description = ""
authors = ["Yoni Baciu"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
langchain = "^0.1.14"
pydantic = "^2.6.4"
langchain-openai = "^0.1.1"
pinecone-client = "^3.2.2"
bs4 = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
